==> nlp/06_team_based_review_2(강사님자료_최신).ipynb <==
## 감성 사전을 만들기(형용사,동사,감탄사 어간 추출 후 파일로 저장)

import pandas as pd
import numpy as np
from konlpy.tag import Okt
okt = Okt()

def extract_stem(words, tokenizer):
    # 문자열을 포함한 다양한 타입에서 NaN 확인
    if pd.isna(words):
        return

    # 전체 텍스트에 대해 한 번에 형태소 분석을 수행
    morphs = tokenizer.morphs(words, stem=True)
    
    # 결과에서 어간만 추출하여 조인
    return ' '.join(morph for morph in morphs if morph)

df_1t = pd.read_csv('./data/review_data_1t_2.csv', encoding='utf-8')
df_1t.head(2)

df_2t = pd.read_csv('./data/review_data_2t_2.csv', encoding='utf-8')
df_2t.head(2)

df_3t = pd.read_csv('./data/review_data_3t_2.csv', encoding='utf-8')
df_3t.head(2)

df_1t.info()

df_2t.info()

df_3t.info()

### 형용사

#### 형용사 중복 제거

# 중복제거전
len(' '.join(df_2t['adj'].dropna().sort_values()).split())

# 중복제거
# 'adj'열 결측치 제거 후 오름차순 정렬후 중복제거 후 공백을 구분자로 한개의 문자열로 만든뒤 공백을 구분자로 리스트 변환
adj_list_1t = ' '.join(df_1t['adj'].dropna().sort_values().drop_duplicates()).split()
# 리스트 내의 중복 제거
adj_list_1t = list(set(adj_list_1t))
print(len(adj_list_1t))

# 중복제거후
adj_list_2t = ' '.join(df_2t['adj'].dropna().sort_values().drop_duplicates()).split()
adj_list_2t = list(set(adj_list_2t))
print(len(adj_list_2t))

# 중복제거후
adj_list_3t = ' '.join(df_3t['adj'].dropna().sort_values().drop_duplicates()).split()
adj_list_3t = list(set(adj_list_3t))
print(len(adj_list_3t))

# 2,3팀 병합 후 중복제거 후 형용사 카운트
tmp=pd.concat([pd.Series(adj_list_2t),pd.Series(adj_list_3t)]).drop_duplicates()
len(tmp)

# 1,2,3팀 병합 후 중복제거 후 형용사 카운트
tmp=pd.concat([pd.Series(adj_list_1t),pd.Series(adj_list_2t),pd.Series(adj_list_3t)]).drop_duplicates()
len(tmp)

#### 형용사 어간 추출

from tqdm import tqdm
tqdm.pandas()
from functools import partial

# extract_stem 함수에 대한 partial 함수 생성
extract_stem_okt = partial(extract_stem, tokenizer=okt)

adj_stem_list_1t = pd.Series(adj_list_1t).progress_map(extract_stem_okt)

# 중복제거
adj_stem_list_1t_unique = adj_stem_list_1t.drop_duplicates()
filtered_adj_stem_list_1t_unique = adj_stem_list_1t_unique[adj_stem_list_1t_unique.map(lambda x:len(x) > 1)]
filtered_adj_stem_list_1t = filtered_adj_stem_list_1t_unique.sort_values().reset_index(drop=True)
filtered_adj_stem_list_1t

adj_stem_list_2t = pd.Series(adj_list_2t).progress_map(extract_stem_okt)

# 중복제거
adj_stem_list_2t_unique = adj_stem_list_2t.drop_duplicates()
filtered_adj_stem_list_2t_unique = adj_stem_list_2t_unique[adj_stem_list_2t_unique.map(lambda x:len(x) > 1)]
filtered_adj_stem_list_2t = filtered_adj_stem_list_2t_unique.sort_values().reset_index(drop=True)
filtered_adj_stem_list_2t

adj_stem_list_3t = pd.Series(adj_list_3t).progress_map(extract_stem_okt)

# 중복제거
adj_stem_list_3t_unique = adj_stem_list_3t.drop_duplicates()
filtered_adj_stem_list_3t_unique = adj_stem_list_3t_unique[adj_stem_list_3t_unique.map(lambda x:len(x) > 1)]
filtered_adj_stem_list_3t = filtered_adj_stem_list_3t_unique.sort_values().reset_index(drop=True)
filtered_adj_stem_list_3t

#### 형용사 어간 파일 저장

tmp = pd.DataFrame(filtered_adj_stem_list_1t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_1t_adj.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(filtered_adj_stem_list_2t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_2t_adj.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(filtered_adj_stem_list_3t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_3t_adj.csv', index=False, encoding='cp949')

tmp = pd.read_csv('./data/sentimental_1t_adj.csv', encoding='cp949')
tmp.head()

### 동사

#### 동사 중복 제거

# 중복제거후
verb_list_1t = ' '.join(df_1t['verb'].dropna().sort_values().drop_duplicates()).split()
verb_list_1t = list(set(verb_list_1t))
print(len(verb_list_1t))

# 중복제거후
verb_list_2t = ' '.join(df_2t['verb'].dropna().sort_values().drop_duplicates()).split()
verb_list_2t = list(set(verb_list_2t))
print(len(adj_list_2t))

# 중복제거후
verb_list_3t = ' '.join(df_3t['verb'].dropna().sort_values().drop_duplicates()).split()
verb_list_3t = list(set(verb_list_3t))
print(len(verb_list_3t))

#### 동사 어간 추출

verb_stem_list_1t = pd.Series(verb_list_1t).progress_map(extract_stem_okt)

# 중복제거
verb_stem_list_1t_unique = verb_stem_list_1t.drop_duplicates()
filtered_verb_stem_list_1t_unique = verb_stem_list_1t_unique[verb_stem_list_1t_unique.map(lambda x:len(x) > 1)]
filtered_verb_stem_list_1t = filtered_verb_stem_list_1t_unique.sort_values().reset_index(drop=True)
filtered_verb_stem_list_1t

from tqdm import tqdm
tqdm.pandas()
verb_stem_list_2t = pd.Series(verb_list_2t).progress_map(extract_stem_okt)

# 중복제거
verb_stem_list_2t_unique = verb_stem_list_2t.drop_duplicates()
filtered_verb_stem_list_2t_unique = verb_stem_list_2t_unique[verb_stem_list_2t_unique.map(lambda x:len(x) > 1)]
filtered_verb_stem_list_2t = filtered_verb_stem_list_2t_unique.sort_values().reset_index(drop=True)
filtered_verb_stem_list_2t

from tqdm import tqdm
tqdm.pandas()
verb_stem_list_3t = pd.Series(verb_list_3t).progress_map(extract_stem_okt)

# 중복제거
verb_stem_list_3t_unique = verb_stem_list_3t.drop_duplicates()
filtered_verb_stem_list_3t_unique = verb_stem_list_3t_unique[verb_stem_list_3t_unique.map(lambda x:len(x) > 1)]
filtered_verb_stem_list_3t = filtered_verb_stem_list_3t_unique.sort_values().reset_index(drop=True)
filtered_verb_stem_list_3t

#### 동사 어간 파일 저장

tmp = pd.DataFrame(filtered_verb_stem_list_1t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_1t_verb.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(filtered_verb_stem_list_1t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_2t_verb.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(filtered_verb_stem_list_1t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_3t_verb.csv', index=False, encoding='cp949')

tmp = pd.read_csv('./data/sentimental_1t_verb.csv', encoding='cp949')
tmp.head()

### 감탄사

#### 감탄사 중복 제거

# 중복제거후
exclamation_list_1t = ' '.join(df_1t['exclamation'].dropna().sort_values().drop_duplicates()).split()
exclamation_list_1t = list(set(exclamation_list_1t))
print(len(exclamation_list_1t))

# 중복제거후
exclamation_list_2t = ' '.join(df_2t['exclamation'].dropna().sort_values().drop_duplicates()).split()
exclamation_list_2t = list(set(exclamation_list_2t))
print(len(exclamation_list_2t))

# 중복제거후
exclamation_list_3t = ' '.join(df_3t['exclamation'].dropna().sort_values().drop_duplicates()).split()
exclamation_list_3t = list(set(exclamation_list_3t))
print(len(exclamation_list_3t))

#### 감탄사 파일저장

tmp = pd.DataFrame(exclamation_list_1t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_1t_exclamation.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(exclamation_list_2t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_2t_exclamation.csv', index=False, encoding='cp949')

tmp = pd.DataFrame(exclamation_list_3t,columns=['단어'])
tmp['점수'] = ''
tmp.to_csv('./data/sentimental_3t_exclamation.csv', index=False, encoding='cp949')

tmp = pd.read_csv('./data/sentimental_1t_exclamation.csv', encoding='cp949')
tmp.head()
