==> nlp/06_team_based_review_2.ipynb <==
# import

import pandas as pd
import numpy as np

# 파일 열기

df_2t = pd.read_csv('data/review_data_2t_2.csv', encoding='utf-8')
df_2t.head(3)

df_2t['noun'].unique()

df_2t.info()

df_2t.isnull().sum()

df_3t = pd.read_csv('data/review_data_3t_2.csv', encoding='utf-8')
df_3t.head(3)

df_3t.info()

df_1t = pd.read_csv('data/review_data_1t_2.csv', encoding='utf-8')
df_1t.head(3)

df_2t['adj'].isnull().sum()

df_2t['adj'].dropna().sort_values()

# 중복 제거 전
len(' '.join(df_2t['adj'].dropna().sort_values()).split())

# 중복 제거 후
adj_list = ' '.join(df_2t['adj'].dropna().sort_values().drop_duplicates()).split()
print(len(adj_list))

len(list(set(adj_list)))

# 중복 제거 후 최종 (2팀)
adj_list_2t = ' '.join(df_2t['adj'].dropna().sort_values().drop_duplicates()).split()
adj_list_2t = list(set(adj_list_2t))  # set로 바꾸면 중복허용이 안되기 때문에
print(len(adj_list_2t))

# 중복 제거 후 최종 (3팀)
adj_list_3t = ' '.join(df_3t['adj'].dropna().sort_values().drop_duplicates()).split()
adj_list_3t = list(set(adj_list_3t))  # set로 바꾸면 중복허용이 안되기 때문에
print(len(adj_list_3t))

# 중복 제거 후 최종 (1팀)
adj_list_1t = ' '.join(df_1t['adj'].dropna().sort_values().drop_duplicates()).split()
adj_list_1t = list(set(adj_list_1t))  # set로 바꾸면 중복허용이 안되기 때문에
print(len(adj_list_1t))

series_1t = pd.Series(adj_list_1t)
series_2t = pd.Series(adj_list_2t)
series_3t = pd.Series(adj_list_3t)

tmp = pd.concat([series_2t, series_3t]).drop_duplicates()

len(tmp)

sorted(adj_list_2t)

## 정규화

from konlpy.tag import Okt

def extract_norm(words):
    if pd.isna(words):
        return words
    
    okt = Okt()
    norm = []
    for word in words:
        # 정규화
        norm = okt.normalize(words)
    return norm

## 어간 추출

# 어간 추출
def extract_stem(words):
    if pd.isna(words):
        return words
    
    okt = Okt()
    stems = []
    for word in words:
        # 어간 추출
        stems = okt.morphs(words,stem=True)
    return stems

from tqdm import tqdm
tqdm.pandas()

## 함수적용

# 정규화
adj_norm_list_2t = pd.Series(adj_list_2t).progress_map(extract_norm)

len(adj_norm_list_2t)

# 어간 추출
adj_stem_list_2t = pd.Series(adj_list_2t).progress_map(extract_stem)

len(adj_norm_list_2t)

adj_norm_list_2t.sort_values()

# 함수 적용 후 중복 제거
adj_norm_list_2t.sort_values().drop_duplicates()

extract_stem('좋지 않다')

extract_stem('좋지')

for word in sorted(adj_list_2t)[:20]:
    print(extract_stem(word))

## 파생열 추가

# NaN값 때문에 astype('str')
df_2t['adj_norm'] = df_2t['adj'].astype('str').progress_map(extract_norm)

df_2t['adj_stem'] = df_2t['adj'].astype('str').progress_map(extract_stem)

df_2t

df_2t.to_csv('data/test.csv', index=False, encoding='utf-8-sig')

df_2t_3 = pd.read_csv('data/review_data_2t_3.csv', encoding='utf-8')
df_2t_3

# adj_norm_list_3t = pd.Series(adj_list_3t).progress_map(extract_stem)

# len(adj_norm_list_3t)

# adj_norm_list_1t = pd.Series(adj_list_1t).progress_map(extract_stem)

# len(adj_norm_list_1t)
