==> nlp/07_team_based_review_3(강사님자료).ipynb <==
## 형용사, 동사 어간추출

import pandas as pd
import numpy as np
from konlpy.tag import Okt
okt = Okt()

def extract_stem(words, tokenizer):
    # 문자열을 포함한 다양한 타입에서 NaN 확인
    if pd.isna(words):
        return

    # 전체 텍스트에 대해 한 번에 형태소 분석을 수행
    morphs = tokenizer.morphs(words, stem=True)
    
    # 결과에서 어간만 추출하여 조인
    return ' '.join(morph for morph in morphs if morph)

extract_stem(df_2t.loc[2,'adj'],okt)

df_1t = pd.read_csv('./data/review_data_1t_2.csv', encoding='utf-8')
df_1t.head(2)

df_2t = pd.read_csv('./data/review_data_2t_2.csv', encoding='utf-8')
df_2t.head(3)

df_3t = pd.read_csv('./data/review_data_3t_2.csv', encoding='utf-8')
df_3t.head(2)

### 형용사 어간 추출하여 파생열 만들기

from tqdm import tqdm
tqdm.pandas()
from functools import partial

# extract_stem 함수에 대한 partial 함수 생성
extract_stem_okt = partial(extract_stem, tokenizer=okt)

df_1t['adj_stem'] = df_1t['adj'].progress_map(extract_stem_okt)

df_2t['adj_stem'] = df_2t['adj'].progress_map(extract_stem_okt)

df_3t['adj_stem'] = df_3t['adj'].progress_map(extract_stem_okt)

### 동사 어간 추출하여 파생열 만들기

df_1t['verb_stem'] = df_1t['verb'].progress_map(extract_stem_okt)

df_2t['verb_stem'] = df_2t['verb'].progress_map(extract_stem_okt)

df_3t['verb_stem'] = df_3t['verb'].progress_map(extract_stem_okt)

df_2t[['adj','adj_stem']]

df_1t.to_csv('./data/review_data_1t_3.csv',index=False)

df_2t.to_csv('./data/review_data_2t_3.csv',index=False)

df_3t.to_csv('./data/review_data_3t_3.csv',index=False)

tmp = pd.read_csv('./data/review_data_2t_3.csv')

tmp
