==> nlp/08_team_based_review_4(강사님자료).ipynb <==
## 감성사전 참조하해서 감성 점수 반영하기

import pandas as pd
import numpy as np

from tqdm import tqdm
tqdm.pandas()

# Review 파일
df_2t = pd.read_csv('./data/review_data_2t_3.csv', encoding='utf-8')

df_adj_sentiment_2t =  pd.read_csv('./data/sentimental/sentimental_2t_adj.csv', encoding='cp949')

# 감성 점수 계산 함수
def calculate_sentiment_score(text):
    if pd.isna(text):
        return 0
    # 감성사전을 사전 형태로 변환
    score_dict = dict(zip(df_adj_sentiment_2t['단어'], df_adj_sentiment_2t['점수']))
    # 텍스트를 단어로 분리
    words = text.split()
    # 각 단어의 점수를 누적하여 합산
    score = sum(score_dict.get(word, 0) for word in words)
    
    # score가 NaN인 경우 0으로 반환
    if np.isnan(score):
        return 0
    else:
        return score
    

calculate_sentiment_score(df_2t.loc[15,'adj_stem'])

df_2t

#### 각 텍스트의 감성 점수를 계산하여 새로운 열에 저장

df_1t['adj_score'] = df_1t['adj_stem'].progress_map(calculate_sentiment_score)
df_1t['verb_score'] = df_1t['verb_stem'].progress_map(calculate_sentiment_score)
df_1t['exclamation_score'] = df_1t['exclamation_stem'].progress_map(calculate_sentiment_score)

df_1t['score'] = df_1t['adj_score'] + df_1t['verb_score'] + df_1t['exclamation_score']

df_2t['adj_score'] = df_2t['adj_stem'].progress_map(calculate_sentiment_score)
df_2t['verb_score'] = df_2t['verb_stem'].progress_map(calculate_sentiment_score)
df_2t['exclamation_score'] = df_2t['exclamation'].progress_map(calculate_sentiment_score)

df_2t['score'] = df_2t['adj_score'] + df_2t['verb_score'] + df_2t['exclamation_score']

# df_2t.to_csv('data/sentimental/text.csv', index=False, encoding='utf-8-sig')

df_2t

df_2t.shape

df_2t[df_2t['score'] > 0 ].shape

df_2t[df_2t['score'] == 0 ].shape

df_2t[df_2t['score'] < 0 ].shape

df_3t['adj_score'] = df_3t['adj_stem'].progress_map(calculate_sentiment_score)
df_3t['verb_score'] = df_3t['verb_stem'].progress_map(calculate_sentiment_score)
df_3t['exclamation_score'] = df_3t['exclamation_stem'].progress_map(calculate_sentiment_score)

df_3t['score'] = df_3t['adj_score'] + df_3t['verb_score'] + df_3t['exclamation_score']
